Separated listener from trait
