+++ base/src/main/java/me/bristermitten/privatemines/PrivateMines.java
@@ -7,6 +7,7 @@ import me.bristermitten.privatemines.commands.PrivateMinesCommand;
 import me.bristermitten.privatemines.config.PMConfig;
 import me.bristermitten.privatemines.config.menu.MenuConfig;
 import me.bristermitten.privatemines.data.MineSchematic;
+import me.bristermitten.privatemines.data.citizens.autosell.AutoSellListener;
 import me.bristermitten.privatemines.data.citizens.autosell.AutoSellNPCTrait;
 import me.bristermitten.privatemines.data.citizens.ultraprisoncore.UltraPrisonCoreNPCTrait;
 import me.bristermitten.privatemines.listeners.UserJoinEvent;
@@ -169,8 +170,9 @@ public final class PrivateMines extends JavaPlugin {
 
         Bukkit.getPluginManager().registerEvents(new UserJoinEvent(), this);
         Bukkit.getPluginManager().registerEvents(new UserLeaveEvent(), this);
-        Bukkit.getPluginManager().registerEvents(new AutoSellNPCTrait(), this);
+        // Bukkit.getPluginManager().registerEvents(new AutoSellNPCTrait(), this);
         Bukkit.getPluginManager().registerEvents(new UltraPrisonCoreNPCTrait(), this);
+        Bukkit.getPluginManager().registerEvents(new AutoSellListener(), this);
 
 //        if (getConfig().getBoolean("autosell-enabled") &&
 //                getConfig().getBoolean("ultraprisoncore-enabled")) {
